Starting to remove all std usage in lexer and parser
