+++ compiler-cpp/f-lang_lib/f-lang_lib.vcxproj
@@ -37,6 +37,7 @@
     <ClInclude Include="..\sources\fstd\memory\array.hpp" />
     <ClInclude Include="..\sources\fstd\memory\bucket.hpp" />
     <ClInclude Include="..\sources\fstd\memory\bucket_array.hpp" />
+    <ClInclude Include="..\sources\fstd\memory\hash_table.hpp" />
     <ClInclude Include="..\sources\fstd\os\windows\console.hpp" />
     <ClInclude Include="..\sources\fstd\platform.hpp" />
     <ClInclude Include="..\sources\fstd\stream\memory_stream.hpp" />
