+++ compiler-cpp/sources/fstd/memory/bucket.hpp
@@ -4,11 +4,12 @@ namespace fstd
 {
 	namespace memory
 	{
-		template<typename Type, size_t size>
+		template<typename Type, size_t capacity>
 		class Bucket
 		{
 			constexpr size_t	size = size;
-			Type				m_elements[size];
+			Type				m_elements[capacity];
+			size_t				m_size;
 		};
 	}
 }
